python/legacy/services/zk-pilot/pyproject.toml
[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[project]
name = "zk-pilot"
version = "0.1.0"
description = "Pilot bootstrap service — logical instance management, token generation, and NATS bootstrap"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "asyncpg>=0.29",
    "nats-py>=2.6",
    "betterproto>=2.0.0b6",
    "pydantic>=2.0",
    "loguru>=0.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
]

[project.scripts]
zk-pilot = "zk_pilot.main:run"

[tool.hatch.build.targets.wheel]
packages = ["src/zk_pilot"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
asyncio_default_test_loop_scope = "session"

[tool.ruff]
line-length = 100

[dependency-groups]
dev = [
    "httpx>=0.28.1",
]
